requirements.txt
ollama
discord.py
python-dotenv